Reenable Tracker reactivity over `hasPermission`
